Port more tests to munit
