+++ prometheus-metrics/src/test/scala/org/http4s/metrics/prometheus/PrometheusServerMetricsSpec.scala
@@ -17,6 +17,7 @@ import org.http4s.testing.Http4sLegacyMatchersIO
 import org.specs2.execute.AsResult
 
 class PrometheusServerMetricsSpec extends Http4sSpec with Http4sLegacyMatchersIO {
+
   private val testRoutes = HttpRoutes.of[IO](stub)
 
   "A http routes with a prometheus metrics middleware" should {
