+++ testing/src/test/scala/org/http4s/Http4sSuite.scala
@@ -7,6 +7,10 @@
 package org.http4s
 
 import cats.effect.Blocker
+import cats.effect.IO
+import cats.syntax.all._
+import fs2._
+import fs2.text.utf8Decode
 import org.http4s.internal.threads.newBlockingPool
 import munit._
 
@@ -15,6 +19,22 @@ import munit._
 trait Http4sSuite extends CatsEffectSuite with DisciplineSuite with munit.ScalaCheckEffectSuite {
 
   val testBlocker: Blocker = Http4sSpec.TestBlocker
+
+  implicit class ParseResultSyntax[A](self: ParseResult[A]) {
+    def yolo: A = self.valueOr(e => sys.error(e.toString))
+  }
+
+  def writeToString[A](a: A)(implicit W: EntityEncoder[IO, A]): IO[String] =
+    Stream
+      .emit(W.toEntity(a))
+      .covary[IO]
+      .flatMap(_.body)
+      .through(utf8Decode)
+      .foldMonoid
+      .compile
+      .last
+      .map(_.getOrElse(""))
+
 }
 
 object Http4sSuite {
